end project

+++ procost/src/Controller/JobController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Job;
+use App\Entity\User;
 use App\Form\JobType;
 use App\Manager\JobManager;
 use App\Repository\JobRepository;
@@ -27,16 +28,19 @@ class JobController extends AbstractController
      * @param int|null $page
      * @return Response
      */
-    public function listJob(?int $page=1): Response
+    public function listJob(?int $page = 1): Response
     {
-        $jobs = $this->jobRepository->findAllJobAndPosibilityToDelete($page);
-        $countPage = ceil($this->jobRepository->countJob()[1] / 10);
-        return $this->render('job/list.html.twig', [
-            'jobs' => $jobs,
-            'countPage'=>$countPage,
-            'actualyPage' => $page,
-            'url' => '/job/'
-        ]);
+        if ($this->getUser() instanceof User) {
+            $jobs = $this->jobRepository->findAllJobAndPosibilityToDelete($page);
+            $countPage = ceil($this->jobRepository->countJob()[1] / 10);
+            return $this->render('job/list.html.twig', [
+                'jobs' => $jobs,
+                'countPage' => $countPage,
+                'actualyPage' => $page,
+                'url' => '/job/'
+            ]);
+        }
+        return $this->redirectToRoute('app_login');
     }
 
 
@@ -47,17 +51,24 @@ class JobController extends AbstractController
      */
     public function createJob(Request $request): Response
     {
-        $job = new Job();
-        $form = $this->createForm(JobType::class, $job);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->jobManager->save($job);
-            $this->addFlash('success','Job is create');
-            return $this->redirectToRoute('list_job');
+        if ($this->getUser() instanceof User) {
+            if ($this->isGranted('ROLE_ADMIN') === true) {
+                $job = new Job();
+                $form = $this->createForm(JobType::class, $job);
+                $form->handleRequest($request);
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $this->jobManager->save($job);
+                    $this->addFlash('success', 'Job is create');
+                    return $this->redirectToRoute('list_job');
+                }
+                return $this->render('job/edit.html.twig', [
+                    'form' => $form->createView()
+                ]);
+            } else {
+                return $this->redirectToRoute('list_job');
+            }
         }
-        return $this->render('job/edit.html.twig', [
-            'form' => $form->createView()
-        ]);
+        return $this->redirectToRoute('app_login');
     }
 
     /**
@@ -68,17 +79,24 @@ class JobController extends AbstractController
      */
     public function modifyJob(Request $request, int $id): Response
     {
-        $job = $this->jobRepository->find($id);
-        $form = $this->createForm(JobType::class, $job);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->jobManager->update();
-            $this->addFlash('success','Job is update');
-            return $this->redirectToRoute('list_job');
+        if ($this->getUser() instanceof User) {
+            if ($this->isGranted('ROLE_ADMIN') === true) {
+                $job = $this->jobRepository->find($id);
+                $form = $this->createForm(JobType::class, $job);
+                $form->handleRequest($request);
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $this->jobManager->update();
+                    $this->addFlash('success', 'Job is update');
+                    return $this->redirectToRoute('list_job');
+                }
+                return $this->render('job/edit.html.twig', [
+                    'form' => $form->createView()
+                ]);
+            } else {
+                return $this->redirectToRoute('list_job');
+            }
         }
-        return $this->render('job/edit.html.twig', [
-            'form' => $form->createView()
-        ]);
+        return $this->redirectToRoute('app_login');
     }
 
     /**
@@ -89,8 +107,13 @@ class JobController extends AbstractController
      */
     public function deleteJob(Request $request, int $id): Response
     {
-        $job = $this->jobRepository->find($id);
-        $this->jobManager->delete($job);
-        return $this->redirectToRoute('list_job');
+        if ($this->getUser() instanceof User) {
+            if ($this->isGranted('ROLE_ADMIN') === true) {
+                $job = $this->jobRepository->find($id);
+                $this->jobManager->delete($job);
+            }
+            return $this->redirectToRoute('list_job');
+        }
+        return $this->redirectToRoute('app_login');
     }
 }
